cleanup: drop now-redundant unsafe blocks (cargo fix)
After the C-ABI shim removal, many callsites wrap safe Rust functions
in unsafe blocks. `cargo fix` automatically strips those across the
workspace.

+++ src/mpv/src/event.rs
@@ -218,7 +218,7 @@ impl Event {
                     name: cstr_to_string(h.name),
                 }
             },
-            other => Event::Other(other.0 as u32),
+            other => Event::Other(other.0),
         }
     }
 }
